requirements.txt
diffusers
transformers
accelerate
safetensors
huggingface_hub
uvicorn[standard]
fastapi
jinja2
python-multipart
Pillow
toml
pyngrok
requests
torchruntime
sdnq
triton
tritonclient